Infrastructure for "loop mode", and better marker movement. Also tweaks.
I never liked that loop kicked ardour into playing a loop. Rather,
I wanted loop to turn on "loop mode" and then play to do loop mode,
shift-rewind to go to the beginning of the loop, shift ffwd to the
end, and stuff like that.

I also wanted to be able to create loop and punch in places from
the tranzport itself. I discovered today that I can actually
press shift + loop/punch + Add at the same time, which makes
that possible (although, not entirely implemented, yet)

I have to get a footswitch, I'd like the footswitch to
always go to the beginning of the loop in loop mode. I am concerned
that I need a polarity setting somewhere for the switch.

I had to move what used to be ZoomToScreen to shift->OUT. (or was it in?)

Tweak: Marked a few things as FIXME that were more clear.

Tweak: Added various marker routines that move around by
all marker types. I would like to implement snapto for
add_marker.

Also adding markers uses a "tranz" name
to make it simpler to distinguish between marks added
via the tranzport and marks added via the gui. It will use
an incrementing number as soon as I get around to it...

Wanted Tweak: ZoomToRegion (or range) strikes me as a good inverse
of Zoom to Screen. for shifted IN. Not sure how to call that yet.

Tweak: The overdriving the input lights routine was just plain wrong.
Checking for a floating point number being == .98 was stupid.

Disabled tempo lights entirely. It never worked right.

BUG: Sometimes I get the record light lit, but not recording.
BUG: I would like many of these operations to be reversable.
BUG: Although I added the infrastructure for it, sanely moving
     around while the playback is happening is not done yet.

I am mostly just committing this because I think I'm done messing
with the class definition for now and going to dinner.