Adding loop and punch ranges now works. A few nits, however.
It is even reversible. It seems to cause some display problems
in the GUI when undone, and sometimes we end up with a loop marker
that is not active, nor shown in the locations menu.

The keystrokes required to do this are a little finicky, however.

Loop->Add is easy
but shift->loop->add is problematic