Alphatrack: Now with More BLING!!
The Bling class has an interesting history. It was the first thing
I'd written to test the tranzport, and was key to me realizing that
libusb was not going to do the job.

I never had a place to put it. I have lost it twice. Now that
I need to abuse the Alphatrack kernel driver I need it or something
like it again. (and it might as well be fun)

AND THIS TIME IT ALL GOES IN GIT.

All I did was ressurect the outline of the code and make it compile
again. I had done something with dynamic bitsets as best as I recall
or some sort of glue template that was wizardly.