Convinced the alphatrack driver to compile once again.
This is still non-working code.

I have to merge back in the fader stuff from the old surfaces branch
and much other code, but the tranzport and alphatrack use a lot of
very similar code and I wanted to get the two closer to being in sync
again....

That branch contained a failed attempt to abstract out the common code,
and this contains a lot of copy/pasted stuff that could one day be
abstracted out, but right now, I don't care.